pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gymnasium_env"]

[project]
name = "gymnasium_env"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "gymnasium",
  "numpy>=1.24.0",
  "pygame>=2.1.3",
  "pre-commit",
  "requests>=2.32.5",
  "streetview>=0.0.16",
  "streetlevel>=0.12.4",
  "pydantic>=2.11.7",
  "folium",
  "pillow>=11.3.0",
  "openai>=1.41.0",
  "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
